+++ system/handlers/error/handler.php
@@ -74,7 +74,7 @@
         Main::set(PAGE,      1);
         Main::set(PAGECOUNT, 1);
 
-        // transfer the handling to the Themes class 
+        // transfer the handling to the Themes class
         Themes::run();
 
         // we handled this page